More extensive weather refactor
